// client/src/types/IEmployeeData.ts
export default interface IEmployeeData {
	epfNo: number;
	initials: string | any;
	firstName: string | any;
	lastName: string | any;
	gender: string | any;
	dob: string | any;
	address: string | any;
	districtId: number;
	provinceId: number;
	contactNo: string | any;
	personalEmail: string | any;
	gsuitEmail: string | any;
	nicNo: string | any;
	nicIssuedDate: string | any;
	passportNo: string | any;
	passExpireDate: string | any;
	licenseNo: string | any;
	licenseIssuedDate: string | any;
	licenseExpireDate: string | any;
	contactPerson: string | any;
	cpRelationship: string | any;
	cpAddress: string | any;
	cpTelephone: string | any;
	cpStatus: string | any;
	cpCivilStatus: string | any;
	cpReligion: string | any;
	appointmentDate: string | any;
	contractStart: string | any;
	contractEnd: string | any;
	locationId: string | any;
	empTypeId: string | any;
	empCatId: string | any;
	designationId: string | any;
	divisionId: string | any;
}